refactor: abstract getDocumentsForCategory query
