+++ src/apis/fetchers/document/get-document/query.ts
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query'
 import { useSession } from 'next-auth/react'
 import { getDocument } from './fetcher'
 
-export const GET_CATEGORY_KEY = 'document'
+export const GET_DOCUMENT_KEY = 'document'
 
 interface Params {
   documentId: number
@@ -14,7 +14,7 @@ export const useGetDocument = ({ documentId }: Params) => {
   const { data: session } = useSession()
 
   return useQuery({
-    queryKey: ['document', documentId],
+    queryKey: [GET_DOCUMENT_KEY, documentId],
     queryFn: () =>
       getDocument({
         accessToken: session?.user.accessToken || '',
