+++ src/app/(routes)/(tabs)/repository/[categoryId]/components/modify-document-name-modal.tsx
@@ -5,7 +5,7 @@ import icons from '@/constants/icons'
 import { Button } from '@/components/ui/button'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useSession } from 'next-auth/react'
-import { Document } from '@/apis/fetchers/document/get-documents-for-category'
+import { Document } from '@/apis/fetchers/document/get-documents-for-category/fetcher'
 import { updateDocumentName } from '@/apis/fetchers/document/update-document-name'
 import { useParams } from 'next/navigation'
 import { SortOption } from './document-list'
